+++ sql/schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE profiles (
     -- Public settings, directly modified by user
     receive_highlights   BOOLEAN DEFAULT TRUE,
     timezone             TEXT DEFAULT NULL,
-    hl_timeout           BIGINT CHECK (hl_timeout >= 1 AND hl_timeout <= 5) DEFAULT 1,
+    hl_timeout           INT CHECK (hl_timeout >= 1 AND hl_timeout <= 5) DEFAULT 1,
     default_ephemeral    BOOLEAN DEFAULT FALSE,
     silence_hl           BOOLEAN DEFAULT FALSE,
     reminders_in_channel BOOLEAN DEFAULT FALSE,
@@ -42,9 +42,10 @@ CREATE TABLE todos (
 CREATE TABLE starboards (
     guild_id    BIGINT PRIMARY KEY,
     channel     BIGINT,
-    threshold   BIGINT DEFAULT 5,
+    threshold   INT CHECK (threshold >= 1) DEFAULT 5,
+    super_mult  INT CHECK (super_mult >= 1) DEFAULT 1,
     format      VARCHAR(200) DEFAULT '⭐ **{stars}**',
-    max_days    BIGINT CHECK (max_days > 1) DEFAULT 7,
+    max_days    INT CHECK (max_days >= 1) DEFAULT 7,
     emoji       TEXT DEFAULT '⭐',
     ignored     BIGINT[] DEFAULT ARRAY[]::BIGINT[],
     FOREIGN KEY (guild_id) REFERENCES guild_configs (guild_id) ON DELETE CASCADE
@@ -55,7 +56,7 @@ CREATE TABLE stars (
     guild_id             BIGINT NOT NULL,
     message_id           BIGINT NOT NULL,
     channel_id           BIGINT NOT NULL,
-    stars                BIGINT NOT NULL,
+    stars                INT NOT NULL,
     starboard_message_id BIGINT NOT NULL,
     PRIMARY KEY (guild_id, message_id, channel_id),
     FOREIGN KEY (guild_id) REFERENCES starboards (guild_id) ON DELETE CASCADE