+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - Can steal custom emoji from other servers via the `emoji` parameter
   - Can create custom emoji from an image file using the `file` and `name` parameters
 - New command `/unicode` to get the codepoints and names of all inputted characters
+- Added `Super Reaction Multiplier` starboard setting, allowing super reactions to contribute more stars to a message
 
 ## Improvements